killpoco: reduce scope and number of various includes.
