+++ test/UnitWOPI.cpp
@@ -17,7 +17,6 @@
 #include <WopiTestServer.hpp>
 
 #include <wsd/DocumentBroker.hpp>
-#include <wsd/Admin.hpp>
 
 #include <Poco/Net/HTTPRequest.h>
